+++ tests/testthat/test-internal.R
@@ -1,28 +1,24 @@
 context("internals not yet caught")
-library(sp)
 library(sf)
-library(raster)
-library(rgdal)
+library(terra)
 library(elevatr)
 data("pt_df")
-data("sp_big")
+data("sf_big")
 data("lake")
 skip_on_cran()
 #skip_on_os("solaris")
-ll_prj  <- "EPSG:4326"
-aea_prj <- "EPSG:5072"
+ll_prj  <- 4326
+aea_prj <- 5072
 
-sp_sm <- SpatialPoints(coordinates(pt_df),
-                       CRS(SRS_string = ll_prj))
-sp_sm_prj <- spTransform(sp_sm, CRS(SRS_string = aea_prj))
-bad_sp <- SpatialPoints(coordinates(data.frame(x = 1000, y = 1000)),
-                        CRS(SRS_string = ll_prj))
-
-sp_sm <- SpatialPoints(coordinates(pt_df),
-                       CRS(SRS_string = ""))
-spdf_sm <- SpatialPointsDataFrame(sp_sm, data.frame(1:nrow(coordinates(sp_sm))))
-
-rast <- rasterize(coordinates(spdf_sm),raster(spdf_sm))
+sf_sm <- st_as_sf(pt_df, coords = c("x", "y"), crs = ll_prj)
+sf_sm_prj <- st_transform(sf_sm, crs = aea_prj) 
+bad_sf <- st_as_sf(data.frame(x = 1000, y = 1000), coords = c("x", "y"), 
+                   crs = ll_prj)
+rast <- terra::rasterize(st_coordinates(sf_sm),terra::rast(sf_sm))
+sf_sm_na <- sf_sm
+st_crs(sf_sm_na) <- NA
+rast_na <- rast
+crs(rast_na) <- NA
 
 mt_wash <- data.frame(x = -71.3036, y = 44.2700)
 mt_mans <- data.frame(x = -72.8145, y = 44.5438)
@@ -37,41 +33,39 @@ test_that("data frame with more extra columns work", {
 
 test_that("proj_expand works",{
   
-  suppressWarnings({
-  mans_sp <- SpatialPoints(coordinates(data.frame(x = -72.8145, y = 44.5438)),
-                           CRS(SRS_string = ll_prj))
-  mans <- get_elev_raster(locations =  mans_sp, z = 6)
-  mans_exp <- get_elev_raster(locations = mans_sp, z = 6, expand = 2)
+  mans_sf <- st_as_sf(data.frame(x = -72.8145, y = 44.5438), 
+                      coords = c("x","y"),
+                      crs = ll_prj)
+  mans <- get_elev_raster(locations =  mans_sf, z = 6)
+  mans_exp <- get_elev_raster(locations = mans_sf, z = 6, expand = 2)
   
-  origin_sp <- SpatialPoints(coordinates(data.frame(x = 0, y = 0)),
-                             CRS(SRS_string = ll_prj))
-  origins <- get_elev_raster(locations = origin_sp, z = 6)
-  })
+  rast_elev <- get_elev_raster(locations = rast, z = 5)
+  rast_elev_exp <- get_elev_raster(locations = rast, z = 5, exp = 5)
   expect_gt(ncell(mans_exp),ncell(mans))
+  expect_gt(ncell(rast_elev_exp), ncell(rast_elev))
   
+  origin_sf <- st_as_sf(data.frame(x = 0, y = 0), coords = c("x", "y"),
+                        crs = ll_prj)
+  origins <- get_elev_raster(locations = origin_sf, z = 6)
+  #expect_is(origins, "SpatRaster")
   expect_is(origins, "RasterLayer")
 })
 
 test_that("loc_check errors correctly", {
-  
+  empty_rast <- rast(nrow = 1, ncol =1)
   expect_error(get_elev_point(locations = pt_df), 
-               "Please supply a valid crs via locations or prj.")
-  expect_error(get_elev_point(locations = sp_sm), 
-               "Please supply a valid crs via locations or prj.")
-  expect_error(get_elev_point(locations = spdf_sm),
-               "Please supply a valid crs via locations or prj.")
-  expect_error(get_elev_point(locations = rast),
-               "Please supply a valid crs via locations or prj.")
-  expect_error(get_elev_point(locations = raster(sp_sm), prj = ll_prj),
-               "No distinct points, all values NA.")
+               "Please supply a valid sf crs via locations or prj.")
+  expect_error(get_elev_point(locations = rast_na),
+               "Please supply a valid sf crs via locations or prj.")
+  expect_error(get_elev_point(locations = sf_sm_na),
+               "Please supply an sf object with a valid crs.")
 })
 
 test_that("Z of 1 or 0 works in get_tilexy",{
+
+  sf_sm_1 <- get_elev_raster(sf_sm, z = 1, clip = "bbox")
+  sf_sm_0 <- get_elev_raster(sf_sm, z = 0, clip = "bbox")
   
-  suppressWarnings({
-  sp_sm_1 <- get_elev_raster(sp_sm_prj, z = 1, clip = "bbox")
-  sp_sm_0 <- get_elev_raster(sp_sm_prj, z = 0, clip = "bbox")
-  })
-  expect_gt(max(res(sp_sm_1)), 0.27)
-  expect_gt(max(res(sp_sm_0)), 0.54)
+  expect_gt(max(res(sf_sm_1)), 0.27)
+  expect_gt(max(res(sf_sm_0)), 0.54)
 })